re-organized component directories, styling changes, ux redundancies added, button styling updated, color themes updated
