+++ client/src/assets/scss/main.scss
@@ -31,9 +31,10 @@ a {
   font-size: 2rem;
   font-weight: 700;
   text-align: center;
-  margin: 0;
+  margin: 0 auto;
   padding: 0;
   color: white;
+  max-width: 85%;
 }
 
 .main-page-container {
