+++ client/src/assets/scss/stylesheets/buttons.scss
@@ -8,11 +8,11 @@
   align-items: center;
   margin: 0 auto;
   margin-top: 1rem;
-
 }
 
 .user-registration-button,
 .sign-in-button {
+  margin-bottom: 1.5rem;
   padding: 4px;
   font-weight: bold;
   border-radius: 5px;
@@ -58,6 +58,40 @@
   font-weight: bolder;
 }
 
-.sort-button {
-  color: black;
+.sort-button, .pay-cash-button {
+  color: white;
+  background-color: $dark-gray;
+  padding: 0.3rem
+}
+.sort-button, .pay-cash-button {
+  color: white;
+  &:hover {
+    cursor: pointer;
+  }
+  &:active {
+    background-color: black;
+  }
+}
+
+.close-payment-options {
+  padding: 0.2rem;
+  margin-top: 0.5rem;
+}
+
+.admin-link {
+  padding: 1rem;
+  margin: 1rem;
+  background-color: $dark-gray;
+  border-radius: 5px;
+  cursor: pointer;
+}
+.admin-link {
+  min-width: 150px;
+  color: white;
+  &:hover {
+    cursor: pointer;
+  }
+  &:active {
+    background-color: black;
+  }
 }
