+++ client/src/assets/scss/stylesheets/topbar.scss
@@ -6,9 +6,9 @@
   justify-content: space-between;
   align-items: center;
   margin: 0 auto;
-  background-color: $dark-gray;
   color: white;
   padding: 1.5rem;
+  border-bottom: 5px solid $dark-gray;
 }
 
 .top-bar-left {
