+++ client/src/components/authentication/SignInForm.js
@@ -1,6 +1,7 @@
 import React, { useState } from "react"
 import config from "../../config"
-import FormError from "../layout/FormError"
+import FormError from "../errorHandling/FormError"
+import { Link } from "react-router-dom"
 
 const SignInForm = () => {
   const [userPayload, setUserPayload] = useState({ email: "", password: "" })
@@ -96,6 +97,7 @@ const SignInForm = () => {
         </div>
         <div className="sign-in-button-container">
           <input type="submit" className="sign-in-button" value="Sign In" />
+          <Link to="/">Cancel/Go Back</Link> 
         </div>
       </form>
     </div>
